# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bethewave
  description: This is a bethewave server
  version: "1.0.0"
  termsOfService: ''
  contact:
    email: hi@bethewave.vote
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers: 
  - url: https://btw-api-staging-2019.herokuapp.com
    description: Staging
  - url: https://btw-api-2019.herokuapp.com
    description: Preprod
  - url: https://btwapi-18.herokuapp.com
    description: Production
...
components:
  schemas:
...
    PotentialVoter:
      type: object
      properties:
        _id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        socialId:
          type: string
        avatar:
          type: string
        socialNetwork:
          type: object
          properties:
            facebook:
              type: boolean
              default: false
            twitter:
              type: boolean
              default: false
        phone: 
          type: string
        isVoter:
          type: boolean
          default: false
        gender:
          type: string
        firstTypeInfo:
          type: object
        secondTypeInfo:
          type: object
        thirdTypeInfo:
          type: object
        captain:
          type: string
        address:
          type: object
          properties:
            city:
              type: string
            region:
              type: string
            poBox:
              type: string
            streetAddress:
              type: string
            extendedAddress:
              type: string
            postalCode:
              type: string
            country:
              type: string
            contryCode:
              type: string
...
  securitySchemes:
    access_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /user/parseFile:
    post:
      tags:
        - user
      summary: Parse file.
      description: 'parse file with potential voters'
      operationId: parseFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: New potential voter were added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  message:
                    type: string
                    example: New potential voters were added
                  nameFile:
                    type: string
                    example: test.csv
        '400':
          description: Bad file
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: number
                    example: 500
                  message: 
                    type: string
                    example: "{ error:'service_unavailable', errorDescription: 'Something went wrong, please try again later', businessErrorCode: 'BE-105' }"
  /user/showResult:
    get:
      tags:
        - user
      summary: Return potential voters.
      description: 'return potential voters for capitan'
      operationId: showResult
      responses:
        '200':
          description: 'successfull'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotentialVoter'
        '400':
          description: not have potential voters
        default:
          description: Internal server error
      security:
        - access_token: []
  /user/addVoters:
    post:
      tags:
        - user
      summary: Add voters.
      description: 'Add voters to the capitan. Number voters not more than 10. Create a collection of voters and associate with the captain'
      operationId: addVoters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                voters:
                  type: array
                  items:
                    type: string
                    description: 'id potential voters'
      responses:
        '200':
          description: New voter was added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '400':
          description: Incorrect data in voters
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
...